+++ Source/CesiumRuntime/Private/CesiumGltfComponent.cpp
@@ -45,9 +45,9 @@
 #if PHYSICS_INTERFACE_PHYSX
 #include "IPhysXCooking.h"
 #else
+#include "Chaos/AABBTree.h"
 #include "Chaos/CollisionConvexMesh.h"
 #include "Chaos/TriangleMeshImplicitObject.h"
-#include "ChaosDerivedDataUtil.h"
 #endif
 
 #if WITH_EDITOR
@@ -55,6 +55,22 @@
 #endif
 
 using namespace CesiumGltf;
+
+namespace {
+
+// UE4 and UE5 both use single-precision vectors for meshes, but they have
+// different names.
+#if ENGINE_MAJOR_VERSION == 5
+using TMeshVector2 = FVector2f;
+using TMeshVector3 = FVector3f;
+using TMeshVector4 = FVector4f;
+#else
+using TMeshVector2 = FVector2D;
+using TMeshVector3 = FVector;
+using TMeshVector4 = FVector4;
+#endif
+
+} // namespace
 
 static uint32_t nextMaterialId = 0;
 
@@ -112,7 +128,14 @@ uint32_t updateTextureCoordinates(
   size_t textureCoordinateIndex = textureCoordinateMap.size();
   textureCoordinateMap[uvAccessorID] = textureCoordinateIndex;
 
-  CesiumGltf::AccessorView<FVector2D> uvAccessor(model, uvAccessorID);
+  struct A {
+    double x[5];
+  };
+
+  const A* a;
+  using Foo = decltype(a->x[0]);
+
+  CesiumGltf::AccessorView<TMeshVector2> uvAccessor(model, uvAccessorID);
   if (uvAccessor.status() != CesiumGltf::AccessorViewStatus::Valid) {
     return 0;
   }
@@ -124,7 +147,7 @@ uint32_t updateTextureCoordinates(
       if (vertexIndex >= 0 && vertexIndex < uvAccessor.size()) {
         vertex.UVs[textureCoordinateIndex] = uvAccessor[vertexIndex];
       } else {
-        vertex.UVs[textureCoordinateIndex] = FVector2D(0.0f, 0.0f);
+        vertex.UVs[textureCoordinateIndex] = TMeshVector2(0.0f, 0.0f);
       }
     }
   } else {
@@ -133,7 +156,7 @@ uint32_t updateTextureCoordinates(
       if (i >= 0 && i < uvAccessor.size()) {
         vertex.UVs[textureCoordinateIndex] = uvAccessor[i];
       } else {
-        vertex.UVs[textureCoordinateIndex] = FVector2D(0.0f, 0.0f);
+        vertex.UVs[textureCoordinateIndex] = TMeshVector2(0.0f, 0.0f);
       }
     }
   }
@@ -161,7 +184,7 @@ static void mikkGetPosition(
     const int VertIdx) {
   TArray<FStaticMeshBuildVertex>& vertices =
       *reinterpret_cast<TArray<FStaticMeshBuildVertex>*>(Context->m_pUserData);
-  FVector& position = vertices[FaceIdx * 3 + VertIdx].Position;
+  const TMeshVector3& position = vertices[FaceIdx * 3 + VertIdx].Position;
   Position[0] = position.X;
   Position[1] = position.Y;
   Position[2] = position.Z;
@@ -174,7 +197,7 @@ static void mikkGetNormal(
     const int VertIdx) {
   TArray<FStaticMeshBuildVertex>& vertices =
       *reinterpret_cast<TArray<FStaticMeshBuildVertex>*>(Context->m_pUserData);
-  FVector& normal = vertices[FaceIdx * 3 + VertIdx].TangentZ;
+  const TMeshVector3& normal = vertices[FaceIdx * 3 + VertIdx].TangentZ;
   Normal[0] = normal.X;
   Normal[1] = normal.Y;
   Normal[2] = normal.Z;
@@ -187,7 +210,7 @@ static void mikkGetTexCoord(
     const int VertIdx) {
   TArray<FStaticMeshBuildVertex>& vertices =
       *reinterpret_cast<TArray<FStaticMeshBuildVertex>*>(Context->m_pUserData);
-  FVector2D& uv = vertices[FaceIdx * 3 + VertIdx].UVs[0];
+  const TMeshVector2& uv = vertices[FaceIdx * 3 + VertIdx].UVs[0];
   UV[0] = uv.X;
   UV[1] = uv.Y;
 }
@@ -201,9 +224,10 @@ static void mikkSetTSpaceBasic(
   TArray<FStaticMeshBuildVertex>& vertices =
       *reinterpret_cast<TArray<FStaticMeshBuildVertex>*>(Context->m_pUserData);
   FStaticMeshBuildVertex& vertex = vertices[FaceIdx * 3 + VertIdx];
-  vertex.TangentX = FVector(Tangent[0], Tangent[1], Tangent[2]);
+  vertex.TangentX = TMeshVector3(Tangent[0], Tangent[1], Tangent[2]);
   vertex.TangentY =
-      BitangentSign * FVector::CrossProduct(vertex.TangentZ, vertex.TangentX);
+      BitangentSign *
+      TMeshVector3::CrossProduct(vertex.TangentZ, vertex.TangentX);
 }
 
 static void computeTangentSpace(TArray<FStaticMeshBuildVertex>& vertices) {
@@ -232,20 +256,26 @@ static void computeFlatNormals(
     FStaticMeshBuildVertex& v1 = vertices[i + 1];
     FStaticMeshBuildVertex& v2 = vertices[i + 2];
 
-    FVector v01 = v1.Position - v0.Position;
-    FVector v02 = v2.Position - v0.Position;
-    FVector normal = FVector::CrossProduct(v01, v02);
+    TMeshVector3 v01 = v1.Position - v0.Position;
+    TMeshVector3 v02 = v2.Position - v0.Position;
+    TMeshVector3 normal = TMeshVector3::CrossProduct(v01, v02);
 
-    v0.TangentX = v1.TangentX = v2.TangentX = FVector(0.0f);
-    v0.TangentY = v1.TangentY = v2.TangentY = FVector(0.0f);
+    v0.TangentX = v1.TangentX = v2.TangentX = TMeshVector3(0.0f);
+    v0.TangentY = v1.TangentY = v2.TangentY = TMeshVector3(0.0f);
     v0.TangentZ = v1.TangentZ = v2.TangentZ = normal.GetSafeNormal();
   }
 }
 
-#if !PHYSICS_INTERFACE_PHYSX
+#if PHYSICS_INTERFACE_PHYSX
+static void BuildPhysXTriangleMeshes(
+    PxTriangleMesh*& pCollisionMesh,
+    const IPhysXCooking* pPhysXCooking,
+    const TArray<FStaticMeshBuildVertex>& vertexData,
+    const TArray<uint32>& indices);
+#else
 static TSharedPtr<Chaos::FTriangleMeshImplicitObject, ESPMode::ThreadSafe>
 BuildChaosTriangleMeshes(
-    const TArray<FStaticMeshBuildVertex>& vertices,
+    const TArray<FStaticMeshBuildVertex>& vertexData,
     const TArray<uint32>& indices);
 #endif
 
@@ -500,7 +530,7 @@ static void loadPrimitive(
     const glm::dmat4x4& transform,
     const CreatePrimitiveOptions& options,
     const CesiumGltf::Accessor& positionAccessor,
-    const CesiumGltf::AccessorView<FVector>& positionView,
+    const CesiumGltf::AccessorView<TMeshVector3>& positionView,
     const TIndexAccessor& indicesView) {
 
   CESIUM_TRACE("loadPrimitive<T>");
@@ -574,11 +604,12 @@ static void loadPrimitive(
   }
 
   auto normalAccessorIt = primitive.attributes.find("NORMAL");
-  CesiumGltf::AccessorView<FVector> normalAccessor;
+  CesiumGltf::AccessorView<TMeshVector3> normalAccessor;
   bool hasNormals = false;
   if (normalAccessorIt != primitive.attributes.end()) {
     int normalAccessorID = normalAccessorIt->second;
-    normalAccessor = CesiumGltf::AccessorView<FVector>(model, normalAccessorID);
+    normalAccessor =
+        CesiumGltf::AccessorView<TMeshVector3>(model, normalAccessorID);
     hasNormals =
         normalAccessor.status() == CesiumGltf::AccessorViewStatus::Valid;
     if (!hasNormals) {
@@ -614,11 +645,11 @@ static void loadPrimitive(
 
   bool hasTangents = false;
   auto tangentAccessorIt = primitive.attributes.find("TANGENT");
-  CesiumGltf::AccessorView<FVector4> tangentAccessor;
+  CesiumGltf::AccessorView<TMeshVector4> tangentAccessor;
   if (tangentAccessorIt != primitive.attributes.end()) {
     int tangentAccessorID = tangentAccessorIt->second;
     tangentAccessor =
-        CesiumGltf::AccessorView<FVector4>(model, tangentAccessorID);
+        CesiumGltf::AccessorView<TMeshVector4>(model, tangentAccessorID);
     hasTangents =
         tangentAccessor.status() == CesiumGltf::AccessorViewStatus::Valid;
     if (!hasTangents) {
@@ -665,9 +696,15 @@ static void loadPrimitive(
       maxPosition = glm::dvec3(max[0], max[1], max[2]);
     }
 
+#if ENGINE_MAJOR_VERSION >= 5
+    FBox aaBox(
+        FVector3d(minPosition.x, minPosition.y, minPosition.z),
+        FVector3d(maxPosition.x, maxPosition.y, maxPosition.z));
+#else
     FBox aaBox(
         FVector(minPosition.x, minPosition.y, minPosition.z),
         FVector(maxPosition.x, maxPosition.y, maxPosition.z));
+#endif
 
     aaBox.GetCenterAndExtents(
         RenderData->Bounds.Origin,
@@ -719,8 +756,8 @@ static void loadPrimitive(
         FStaticMeshBuildVertex& vertex = StaticMeshBuildVertices[i];
         uint32 vertexIndex = indices[i];
         vertex.Position = positionView[vertexIndex];
-        vertex.UVs[0] = FVector2D(0.0f, 0.0f);
-        vertex.UVs[2] = FVector2D(0.0f, 0.0f);
+        vertex.UVs[0] = TMeshVector2(0.0f, 0.0f);
+        vertex.UVs[2] = TMeshVector2(0.0f, 0.0f);
         RenderData->Bounds.SphereRadius = FMath::Max(
             (vertex.Position - RenderData->Bounds.Origin).Size(),
             RenderData->Bounds.SphereRadius);
@@ -730,8 +767,8 @@ static void loadPrimitive(
       for (int64_t i = 0; i < StaticMeshBuildVertices.Num(); ++i) {
         FStaticMeshBuildVertex& vertex = StaticMeshBuildVertices[i];
         vertex.Position = positionView[i];
-        vertex.UVs[0] = FVector2D(0.0f, 0.0f);
-        vertex.UVs[2] = FVector2D(0.0f, 0.0f);
+        vertex.UVs[0] = TMeshVector2(0.0f, 0.0f);
+        vertex.UVs[2] = TMeshVector2(0.0f, 0.0f);
         RenderData->Bounds.SphereRadius = FMath::Max(
             (vertex.Position - RenderData->Bounds.Origin).Size(),
             RenderData->Bounds.SphereRadius);
@@ -854,16 +891,16 @@ static void loadPrimitive(
       for (int64_t i = 0; i < indices.Num(); ++i) {
         FStaticMeshBuildVertex& vertex = StaticMeshBuildVertices[i];
         uint32 vertexIndex = indices[i];
-        vertex.TangentX = FVector(0.0f, 0.0f, 0.0f);
-        vertex.TangentY = FVector(0.0f, 0.0f, 0.0f);
+        vertex.TangentX = TMeshVector3(0.0f, 0.0f, 0.0f);
+        vertex.TangentY = TMeshVector3(0.0f, 0.0f, 0.0f);
         vertex.TangentZ = normalAccessor[vertexIndex];
       }
     } else {
       CESIUM_TRACE("copy normals");
       for (int64_t i = 0; i < StaticMeshBuildVertices.Num(); ++i) {
         FStaticMeshBuildVertex& vertex = StaticMeshBuildVertices[i];
-        vertex.TangentX = FVector(0.0f, 0.0f, 0.0f);
-        vertex.TangentY = FVector(0.0f, 0.0f, 0.0f);
+        vertex.TangentX = TMeshVector3(0.0f, 0.0f, 0.0f);
+        vertex.TangentY = TMeshVector3(0.0f, 0.0f, 0.0f);
         vertex.TangentZ = normalAccessor[i];
       }
     }
@@ -878,19 +915,21 @@ static void loadPrimitive(
       for (int64_t i = 0; i < indices.Num(); ++i) {
         FStaticMeshBuildVertex& vertex = StaticMeshBuildVertices[i];
         uint32 vertexIndex = indices[i];
-        const FVector4& tangent = tangentAccessor[vertexIndex];
+        const TMeshVector4& tangent = tangentAccessor[vertexIndex];
         vertex.TangentX = tangent;
         vertex.TangentY =
-            FVector::CrossProduct(vertex.TangentZ, vertex.TangentX) * tangent.W;
+            TMeshVector3::CrossProduct(vertex.TangentZ, vertex.TangentX) *
+            tangent.W;
       }
     } else {
       CESIUM_TRACE("copy tangents");
       for (int64_t i = 0; i < StaticMeshBuildVertices.Num(); ++i) {
         FStaticMeshBuildVertex& vertex = StaticMeshBuildVertices[i];
-        const FVector4& tangent = tangentAccessor[i];
+        const TMeshVector4& tangent = tangentAccessor[i];
         vertex.TangentX = tangent;
         vertex.TangentY =
-            FVector::CrossProduct(vertex.TangentZ, vertex.TangentX) * tangent.W;
+            TMeshVector3::CrossProduct(vertex.TangentZ, vertex.TangentX) *
+            tangent.W;
       }
     }
   }
@@ -920,8 +959,13 @@ static void loadPrimitive(
         false);
   }
 
+#if ENGINE_MAJOR_VERSION == 5
+  FStaticMeshSectionArray& Sections = LODResources.Sections;
+#else
   FStaticMeshLODResources::FStaticMeshSectionArray& Sections =
       LODResources.Sections;
+#endif
+
   FStaticMeshSection& section = Sections.AddDefaulted_GetRef();
   section.bEnableCollision = true;
 
@@ -963,7 +1007,9 @@ static void loadPrimitive(
   LODResources.bHasDepthOnlyIndices = false;
   LODResources.bHasReversedIndices = false;
   LODResources.bHasReversedDepthOnlyIndices = false;
+#if ENGINE_MAJOR_VERSION < 5
   LODResources.bHasAdjacencyInfo = false;
+#endif
 
   primitiveResult.pModel = &model;
   primitiveResult.pMeshPrimitive = &primitive;
@@ -975,46 +1021,20 @@ static void loadPrimitive(
 
   primitiveResult.pCollisionMesh = nullptr;
 
-#if PHYSICS_INTERFACE_PHYSX
-  if (options.pMeshOptions->pNodeOptions->pModelOptions->pPhysXCooking) {
-    CESIUM_TRACE("PhysX cook");
-    // TODO: use PhysX interface directly so we don't need to copy the
-    // vertices (it takes a stride parameter).
-    TArray<FVector> vertices;
-    vertices.SetNum(StaticMeshBuildVertices.Num());
-
-    for (size_t i = 0; i < StaticMeshBuildVertices.Num(); ++i) {
-      vertices[i] = StaticMeshBuildVertices[i].Position;
-    }
-
-    TArray<FTriIndices> physicsIndices;
-    physicsIndices.SetNum(indices.Num() / 3);
-
-    for (size_t i = 0; i < indices.Num() / 3; ++i) {
-      physicsIndices[i].v0 = indices[3 * i];
-      physicsIndices[i].v1 = indices[3 * i + 1];
-      physicsIndices[i].v2 = indices[3 * i + 2];
-    }
-
-    options.pMeshOptions->pNodeOptions->pModelOptions->pPhysXCooking
-        ->CreateTriMesh(
-            "PhysXGeneric",
-            EPhysXMeshCookFlags::Default,
-            vertices,
-            physicsIndices,
-            TArray<uint16>(),
-            true,
-            primitiveResult.pCollisionMesh);
-  }
-#else
   if (StaticMeshBuildVertices.Num() != 0 && indices.Num() != 0) {
-    CESIUM_TRACE("Chaos cook");
-    primitiveResult.pCollisionMesh = BuildChaosTriangleMeshes(
-        duplicateVertices,
+#if PHYSICS_INTERFACE_PHYSX
+    CESIUM_TRACE("PhysX cook");
+    BuildPhysXTriangleMeshes(
+        primitiveResult.pCollisionMesh,
+        options.pMeshOptions->pNodeOptions->pModelOptions->pPhysXCooking,
         StaticMeshBuildVertices,
         indices);
-  }
+#else
+    CESIUM_TRACE("Chaos cook");
+    primitiveResult.pCollisionMesh =
+        BuildChaosTriangleMeshes(StaticMeshBuildVertices, indices);
 #endif
+  }
 
   // load primitive metadata
   primitiveResult.Metadata = loadMetadataPrimitive(model, primitive);
@@ -1025,7 +1045,7 @@ static void loadIndexedPrimitive(
     const glm::dmat4x4& transform,
     const CreatePrimitiveOptions& options,
     const CesiumGltf::Accessor& positionAccessor,
-    const CesiumGltf::AccessorView<FVector>& positionView) {
+    const CesiumGltf::AccessorView<TMeshVector3>& positionView) {
 
   const Model& model =
       *options.pMeshOptions->pNodeOptions->pModelOptions->pModel;
@@ -1114,7 +1134,9 @@ static void loadPrimitive(
     return;
   }
 
-  CesiumGltf::AccessorView<FVector> positionView(model, *pPositionAccessor);
+  CesiumGltf::AccessorView<TMeshVector3> positionView(
+      model,
+      *pPositionAccessor);
 
   if (primitive.indices < 0 || primitive.indices >= model.accessors.size()) {
     std::vector<uint32_t> syntheticIndexBuffer(positionView.size());
@@ -1500,11 +1522,17 @@ static void loadPrimitiveGameThreadPart(
   pStaticMesh->NeverStream = true;
 
 #if ENGINE_MAJOR_VERSION == 4 && ENGINE_MINOR_VERSION < 27
+  // UE 4.26 or earlier
   pStaticMesh->bIsBuiltAtRuntime = true;
   pStaticMesh->RenderData =
       TUniquePtr<FStaticMeshRenderData>(loadResult.RenderData);
-#else
+#elif ENGINE_MAJOR_VERSION == 4
+  // UE 4.27 or later
   pStaticMesh->SetIsBuiltAtRuntime(true);
+  pStaticMesh->SetRenderData(
+      TUniquePtr<FStaticMeshRenderData>(loadResult.RenderData));
+#else
+  // UE 5
   pStaticMesh->SetRenderData(
       TUniquePtr<FStaticMeshRenderData>(loadResult.RenderData));
 #endif
@@ -1563,16 +1591,21 @@ static void loadPrimitiveGameThreadPart(
   if (pBaseAsMaterialInstance && !pCesiumData) {
     const FStaticParameterSet& parameters =
         pBaseAsMaterialInstance->GetStaticParameters();
+
+#if ENGINE_MAJOR_VERSION >= 5
+    bool hasLayers = parameters.bHasMaterialLayers;
+#else
     const TArray<FStaticMaterialLayersParameter>& layerParameters =
         parameters.MaterialLayersParameters;
-
     const FStaticMaterialLayersParameter* pCesiumLayers =
         layerParameters.FindByPredicate(
             [](const FStaticMaterialLayersParameter& layerParameter) {
               return layerParameter.ParameterInfo.Name == "Cesium";
             });
+    bool hasLayers = pCesiumLayers != nullptr;
+#endif
 
-    if (pCesiumLayers) {
+    if (hasLayers) {
 #if WITH_EDITOR
       FScopedTransaction transaction(
           FText::FromString("Add Cesium User Data to Material"));
@@ -1624,22 +1657,23 @@ static void loadPrimitiveGameThreadPart(
 #endif
   pStaticMesh->CreateBodySetup();
 
+  UBodySetup* pBodySetup = pMesh->GetBodySetup();
+
   // pMesh->UpdateCollisionFromStaticMesh();
-  pMesh->GetBodySetup()->CollisionTraceFlag =
-      ECollisionTraceFlag::CTF_UseComplexAsSimple;
+  pBodySetup->CollisionTraceFlag = ECollisionTraceFlag::CTF_UseComplexAsSimple;
 
   if (loadResult.pCollisionMesh) {
 #if PHYSICS_INTERFACE_PHYSX
-    pMesh->GetBodySetup()->TriMeshes.Add(loadResult.pCollisionMesh);
+    pBodySetup->TriMeshes.Add(loadResult.pCollisionMesh);
 #else
-    pMesh->GetBodySetup()->ChaosTriMeshes.Add(loadResult.pCollisionMesh);
+    pBodySetup->ChaosTriMeshes.Add(loadResult.pCollisionMesh);
 #endif
   }
 
   // Mark physics meshes created, no matter if we actually have a collision
   // mesh or not. We don't want the editor creating collision meshes itself in
   // the game thread, because that would be slow.
-  pMesh->GetBodySetup()->bCreatedPhysicsMeshes = true;
+  pBodySetup->bCreatedPhysicsMeshes = true;
 
   pMesh->SetMobility(EComponentMobility::Movable);
 
@@ -1759,7 +1793,7 @@ void forEachPrimitiveComponent(UCesiumGltfComponent* pGltf, Func&& f) {
       UMaterialInstanceDynamic* pMaterial =
           Cast<UMaterialInstanceDynamic>(pPrimitive->GetMaterial(0));
 
-      if (pMaterial->IsPendingKillOrUnreachable()) {
+      if (!IsValid(pMaterial)) {
         // Don't try to update the material while it's in the process of being
         // destroyed. This can lead to the render thread freaking out when
         // it's asked to update a parameter for a material that has been
@@ -1902,23 +1936,112 @@ void UCesiumGltfComponent::SetCollisionEnabled(
   }
 }
 
-#if !PHYSICS_INTERFACE_PHYSX
-// This is copied from FChaosDerivedDataCooker::BuildTriangleMeshes in
-// C:\Program Files\Epic
-// Games\UE_4.26\Engine\Source\Runtime\Engine\Private\PhysicsEngine\Experimental\ChaosDerivedData.cpp.
-// We can't use that method directly because it is private. Since we've copied
-// it anyway, we've also modified it to be more efficient for our particular
-// input data.
+#if PHYSICS_INTERFACE_PHYSX
+static void BuildPhysXTriangleMeshes(
+    PxTriangleMesh*& pCollisionMesh,
+    const IPhysXCooking* pPhysXCooking,
+    const TArray<FStaticMeshBuildVertex>& vertexData,
+    const TArray<uint32>& indices) {
+
+  if (pPhysXCooking) {
+    int32 vertexCount = vertexData.Num();
+    int32 triangleCount = indices.Num() / 3;
+
+    // TODO: use PhysX interface directly so we don't need to copy the
+    // vertices (it takes a stride parameter).
+    TArray<FVector> vertices;
+    vertices.SetNum(vertexCount);
+
+    for (size_t i = 0; i < vertexCount; ++i) {
+      vertices[i] = vertexData[i].Position;
+    }
+
+    TArray<FTriIndices> physicsIndices;
+    physicsIndices.SetNum(triangleCount);
+
+    for (size_t i = 0; i < triangleCount; ++i) {
+      physicsIndices[i].v0 = indices[3 * i];
+      physicsIndices[i].v1 = indices[3 * i + 1];
+      physicsIndices[i].v2 = indices[3 * i + 2];
+    }
+
+    pPhysXCooking->CreateTriMesh(
+        "PhysXGeneric",
+        EPhysXMeshCookFlags::Default,
+        vertices,
+        physicsIndices,
+        TArray<uint16>(),
+        true,
+        pCollisionMesh);
+  }
+}
+
+#else
+template <typename TIndex>
+static void fillTriangles(
+    TArray<Chaos::TVector<TIndex, 3>>& triangles,
+    const TArray<FStaticMeshBuildVertex>& vertexData,
+    const TArray<uint32>& indices,
+    int32 triangleCount) {
+
+  triangles.Reserve(triangleCount);
+
+  for (TIndex i = 0; i < static_cast<TIndex>(triangleCount); ++i) {
+    TIndex index0 = 3 * i;
+    triangles.Add(Chaos::TVector<TIndex, 3>(
+        static_cast<TIndex>(indices[index0 + 1]),
+        static_cast<TIndex>(indices[index0]),
+        static_cast<TIndex>(indices[index0 + 2])));
+  }
+}
+
 static TSharedPtr<Chaos::FTriangleMeshImplicitObject, ESPMode::ThreadSafe>
 BuildChaosTriangleMeshes(
-    bool duplicateVertices,
-    const TArray<FStaticMeshBuildVertex>& vertices,
+    const TArray<FStaticMeshBuildVertex>& vertexData,
     const TArray<uint32>& indices) {
-  UE_LOG(
-      LogCesium,
-      Warning,
-      TEXT(
-          "The Chaos physics engine is not currently supported by Cesium for Unreal because functionality required to cook meshes at runtime is not available."));
-  return nullptr;
+
+  int32 vertexCount = vertexData.Num();
+  int32 triangleCount = indices.Num() / 3;
+
+  Chaos::TParticles<Chaos::FRealSingle, 3> vertices;
+  vertices.AddParticles(vertexCount);
+
+  for (int32 i = 0; i < vertexCount; ++i) {
+    vertices.X(i) = vertexData[i].Position;
+  }
+
+  TArray<uint16> materials;
+  materials.SetNum(triangleCount);
+
+  TArray<int32> faceRemap;
+  faceRemap.SetNum(triangleCount);
+
+  for (int32 i = 0; i < triangleCount; ++i) {
+    faceRemap[i] = i;
+  }
+
+  TUniquePtr<TArray<int32>> pFaceRemap = MakeUnique<TArray<int32>>(faceRemap);
+
+  if (vertexCount < TNumericLimits<uint16>::Max()) {
+    TArray<Chaos::TVector<uint16, 3>> triangles;
+    fillTriangles(triangles, vertexData, indices, triangleCount);
+    return MakeShared<Chaos::FTriangleMeshImplicitObject, ESPMode::ThreadSafe>(
+        MoveTemp(vertices),
+        MoveTemp(triangles),
+        MoveTemp(materials),
+        MoveTemp(pFaceRemap),
+        nullptr,
+        false);
+  } else {
+    TArray<Chaos::TVector<int32, 3>> triangles;
+    fillTriangles(triangles, vertexData, indices, triangleCount);
+    return MakeShared<Chaos::FTriangleMeshImplicitObject, ESPMode::ThreadSafe>(
+        MoveTemp(vertices),
+        MoveTemp(triangles),
+        MoveTemp(materials),
+        MoveTemp(pFaceRemap),
+        nullptr,
+        false);
+  }
 }
 #endif